requirements.txt
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cryptography==40.0.2
dataclasses==0.8
Flask==2.0.3
greenlet==2.0.2
idna==3.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.54
typing_extensions==4.1.1
urllib3==1.26.20
waitress==2.0.0
Werkzeug==2.0.3
zipp==3.6.0
